[CS-3836]: Use single stack
