+++ cardstack/src/navigation/tabBarNavigator.tsx
@@ -2,19 +2,11 @@ import {
   BottomTabBarOptions,
   createBottomTabNavigator,
 } from '@react-navigation/bottom-tabs';
-import { NavigationContainer } from '@react-navigation/native';
 import {
   createStackNavigator,
   StackNavigationOptions,
 } from '@react-navigation/stack';
-import React, {
-  createContext,
-  Dispatch,
-  SetStateAction,
-  useContext,
-  useMemo,
-  useState,
-} from 'react';
+import React, { useContext } from 'react';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
 import { TabBarIcon } from '@cardstack/components';
@@ -27,9 +19,7 @@ import {
 import { colors } from '@cardstack/theme';
 import { Device } from '@cardstack/utils';
 
-import { navigationRef } from '@rainbow-me/navigation/Navigation';
 import { expandedPreset, sheetPreset } from '@rainbow-me/navigation/effects';
-import { onNavigationStateChange } from '@rainbow-me/navigation/onNavigationStateChange';
 import RainbowRoutes from '@rainbow-me/navigation/routesNames';
 import ChangeWalletSheet from '@rainbow-me/screens/ChangeWalletSheet';
 import ExpandedAssetSheet from '@rainbow-me/screens/ExpandedAssetSheet';
@@ -39,10 +29,9 @@ import RestoreSheet from '@rainbow-me/screens/RestoreSheet';
 
 import { InitialRouteContext } from '../../../src/context/initialRoute';
 
-import { useCardstackGlobalScreens, useCardstackMainScreens } from './hooks';
-import { linking } from './screens';
+import { useCardstackMainScreens } from './hooks';
 
-import { dismissAndroidKeyboardOnClose, tabLinking } from '.';
+import { dismissAndroidKeyboardOnClose } from '.';
 
 const Tab = createBottomTabNavigator();
 
@@ -121,16 +110,12 @@ const TabNavigator = () => {
 
 const Stack = createStackNavigator();
 
-const StackNavigator = () => {
+export const StackNavigator = () => {
   const initialRoute = useContext(InitialRouteContext) || '';
 
   const cardstackMainScreens = useCardstackMainScreens(Stack);
-  const cardstackGlobalScreens = useCardstackGlobalScreens(Stack);
 
   // TODO: Create a navigator for each flow and split auth/non-auth
-
-  // Remove last item aka LoadingOverlay, to avoid dupe (on iOS)
-  Device.isIOS && cardstackGlobalScreens.pop();
 
   return (
     <Stack.Navigator
@@ -149,7 +134,6 @@ const StackNavigator = () => {
         name={RainbowRoutes.SWIPE_LAYOUT}
       />
       {cardstackMainScreens}
-      {cardstackGlobalScreens}
 
       {
         // Temp rainbow components until migration
@@ -191,41 +175,3 @@ const StackNavigator = () => {
     </Stack.Navigator>
   );
 };
-
-// Temp feature flag context
-interface TabBarContextType {
-  isTabBarEnabled: boolean;
-  setIsTabBarEnabled: Dispatch<SetStateAction<boolean>>;
-}
-
-const TabBarFeatureContext = createContext<TabBarContextType>({
-  isTabBarEnabled: false,
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  setIsTabBarEnabled: () => {},
-});
-
-export const TabBarFeatureProvider: React.FC = ({ children }) => {
-  const [isTabBarEnabled, setIsTabBarEnabled] = useState(true);
-
-  const contextValues = useMemo(
-    () => ({
-      isTabBarEnabled,
-      setIsTabBarEnabled,
-    }),
-    [isTabBarEnabled]
-  );
-
-  return (
-    <NavigationContainer
-      linking={isTabBarEnabled ? tabLinking : linking}
-      onStateChange={onNavigationStateChange}
-      ref={navigationRef}
-    >
-      <TabBarFeatureContext.Provider value={contextValues}>
-        {isTabBarEnabled ? <StackNavigator /> : children}
-      </TabBarFeatureContext.Provider>
-    </NavigationContainer>
-  );
-};
-
-export const useTabBarFlag = () => useContext(TabBarFeatureContext);
